# apps/clankie/openapi.yaml
openapi: 3.1.0
info:
  title: Clankie
  version: "0.2.0"
  description: |
    Local `apps/clankie` service on `http://127.0.0.1:4310`. There is no HTML
    index: `/` is 404 and `GET /health` is the public probe.

    Most other routes take an `Authorization: Bearer <token>` header. Pairing
    redeem/complete use their one-time offer secrets instead. Persistent tokens
    live in the credential broker (Keychain service `bot.clankie.credentials`):

    | Principal | Keychain account | Typical use |
    | --- | --- | --- |
    | Operator | `clankie_operator` | TUI operator reads/writes, pairing, devices, play stop |
    | Captain | `clankie_captain` (also `CLANKIE_CAPTAIN_TOKEN`) | `/operator/v1/dispatch`, `/captain/v1/lanes` |
    | Discord text | `clankie_discord_bridge` | Discord channel turns, presence, person memory |
    | Discord voice | `clankie_discord_voice_bridge` | Voice briefing; voice-lane channel turns |
    | Discord user text | `clankie_discord_user_bridge` | Lab-user channel turns and presence |
    | Discord user voice | `clankie_discord_user_voice_bridge` | Lab-user voice-lane turns |
    | Device | minted at pairing complete | `/v1/devices/self*` |

    Discord-lane routes 403 if the bearer is the console captain token (steer
    source `api`). Import this file in Yaak (Settings → Import/Export) or run
    `apps/clankie/scripts/setup-yaak.py`.
servers:
  - url: http://127.0.0.1:4310
    description: Local service
...
security: []
paths:
...
  /v1/accounts/codex:
    get:
      tags: [Operator]
      operationId: listCodexAccounts
      security: [{ operatorBearer: [] }, { deviceBearer: [] }]
      summary: List registered local Codex homes and observed headroom
      description: Owner-authorized. Reports credential-file presence, never credential contents. Queries current quota without a model turn, with recent rollouts as fallback. Missing or stale usage is unknown.
      responses:
        "200":
          description: Default and registered account labels, paths, rate windows and headroom
        "401":
          description: Authentication required
        "403":
          description: Owner authority required
    post:
      tags: [Operator]
      operationId: updateCodexAccounts
      security: [{ operatorBearer: [] }, { deviceBearer: [] }]
      summary: Register or remove an owner's local Codex home
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [op, home, label]
                  additionalProperties: false
                  properties:
                    op: { type: string, enum: [add] }
                    home: { type: string }
                    label: { type: string }
                - type: object
                  required: [op, label]
                  additionalProperties: false
                  properties:
                    op: { type: string, enum: [remove] }
                    label: { type: string }
      responses:
        "200":
          description: Registration updated; credentials and home are unchanged
        "400":
          description: Invalid or duplicate account
        "401":
          description: Authentication required
        "403":
          description: Owner authority required
...
  /v1/accounts/linear/app:
    post:
      tags: [Operator]
      operationId: connectLinearApp
      summary: Connect a workspace-owned Linear app for worker names and portraits
      description: |
        Verifies an app actor and workspace before replacing the Linear credential.
        Owner operator bearer or active device with terminalControl only. Public
        gateway calls require the encrypted envelope. Client credentials and tokens
        stay in the broker and are never returned or logged. Cache-Control no-store.
        Enable client credentials tokens on the workspace's Linear OAuth application.
      security: [{ operatorBearer: [] }, { deviceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [clientId, clientSecret]
              properties:
                clientId: { type: string, minLength: 1, maxLength: 256 }
                clientSecret: { type: string, minLength: 1, maxLength: 4096, writeOnly: true }
      responses:
        "200":
          description: Verified connection with actor and workspace; no credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountResult" }
        "400":
          description: malformed or provider_rejected; previous connection retained
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountResult" }
        "401": { description: authentication_required }
        "403": { description: forbidden — terminalControl required }
        "503": { description: unavailable }
...
components:
  schemas:
...
    AccountResult:
      type: object
      description: |
        ok true with the operation's fields (flowId, userCode, verificationUri, interval,
        status, authorizeUrl, redirectUri, expiresAt, connection, revoked, manageUrl), or
        ok false with a closed error code. See @clankie/protocol/accounts.
      required: [ok]
      properties:
        ok: { type: boolean }
        error:
          type: string
          enum:
            [
              authentication_required,
              forbidden,
              malformed,
              unconfigured,
              unknown_flow,
              expired,
              denied,
              provider_rejected,
              unavailable,
            ]
...
  securitySchemes:
    swarmWorkerBearer:
      type: http
      scheme: bearer
      description: Runtime's enrolled SWARM_SESSION_CAPABILITY, verified at the grant's selected coordinator; never an operator or provider credential
    workerBearer:
      type: http
      scheme: bearer
      description: Individual short-lived grant issued by POST /v1/worker-grants/; never an operator or provider credential
    operatorBearer:
      type: http
      scheme: bearer
      description: Keychain `clankie_operator`
    captainBearer:
      type: http
      scheme: bearer
      description: Keychain `clankie_captain` / `CLANKIE_CAPTAIN_TOKEN`
    discordTextBearer:
      type: http
      scheme: bearer
      description: Keychain `clankie_discord_bridge` or `clankie_discord_user_bridge`
    discordVoiceBearer:
      type: http
      scheme: bearer
      description: Keychain `clankie_discord_voice_bridge` or `clankie_discord_user_voice_bridge`
    deviceBearer:
      type: http
      scheme: bearer
      description: Device session token from pairing complete